Refactor into multiple pages
